+++ NEWS.md
@@ -9,6 +9,7 @@
 - Add `date_to_datetime()` that converts a date to a date time in the corresponding timezone based on `hub` and returns that date time in UTC timezone. This function is used when the user is using `as_of` parameter to load forecasts from zoltar only.
 - Add `preprocess_visualization_truth()` to generate JSON truth file for covid19 hub visualization, and its corresponding unit tests
 - Add `calc_cramers_dist_equal_space()`, `calc_cramers_dist_equal_space()`, and `calc_cramers_dist_one_model_pair()` to calculate forecast similarities based on the approximation of Cramer's distance.
+- In `score_forecasts()` calculate one-sided quantile coverage denoted quantile_coverage0.xx.
 
 ## covidHubUtils 0.1.5